chore: remove `no-window-during-ssr` and `no-document-during-ssr`
