+++ lib/index.js
@@ -28,11 +28,9 @@ const rules = {
     'valid-api': require('./rules/valid-api'),
     'valid-track': require('./rules/valid-track'),
     'valid-wire': require('./rules/valid-wire'),
-    'no-document-during-ssr': require('./rules/no-document-during-ssr'),
     'no-querying-slots-during-ssr': require('./rules/no-querying-slots-during-ssr'),
     'no-restricted-browser-globals-during-ssr': require('./rules/no-restricted-browser-globals-during-ssr'),
     'no-this-template-during-ssr': require('./rules/no-this-template-during-ssr'),
-    'no-window-during-ssr': require('./rules/no-window-during-ssr'),
 };
 
 module.exports = {
