+++ lib/rules/no-restricted-browser-globals-during-ssr.js
@@ -620,6 +620,7 @@ const prohibitedGlobalVariables = new Set([
     'crossOriginIsolated',
     'customElements',
     'devicePixelRatio',
+    'document',
     'event',
     'external',
     'find',
@@ -816,6 +817,7 @@ const prohibitedGlobalVariables = new Set([
     'webkitSpeechRecognitionEvent',
     'webkitStorageInfo',
     'webkitURL',
+    'window',
 ]);
 
 module.exports = {
